Remove separate options 'external-ccsdv' and 'external-fciv'. Now, there is just an 'external' keyword for the external correction.
Instead, whether the interaction in the T3 * V term is contracted
with V in the high-level (i.e. FCI) 'smaller' space, or the low-level
(i.e. CCSD) 'larger' space, is controlled via an optional argument to the
'externally_correct' function. This argument is 'low_level_coul', and
defaults to 'True'. This means that the default behaviour is equivalent
to the 'external-ccsdv' correction previously. Tests and examples updated.